Update items controller actions to considr logged user

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,28 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
         private readonly ApplicationDbContext _context;
-
-        public ItemController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+        
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+            var model = _context.Items.Where(item => item.User.Id == user.Id).ToList();
 
             return View("Index", model);
         }
@@ -36,7 +43,15 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+            var user = _userManager.GetUserAsync(HttpContext.User).Result;
+
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+
+            if (item.User.Id != user.Id)
+            {
+                return this.Unauthorized();
+            }
+
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
